Added storage checks to filesaver

+++ naturewatch_camera_server/FileSaver.py
@@ -3,6 +3,7 @@ import cv2
 import io
 import logging
 import os
+import datetime
 from subprocess import call
 try:
     import picamera
@@ -31,17 +32,20 @@ class FileSaver(Thread):
         :param image: numpy array image
         :return: filename
         """
-
-        filename = timestamp
-        filename = filename + ".jpg"
-        self.logging.info('saving file')
-        try:
-            cv2.imwrite(os.path.join(self.config["photos_path"], filename), image)
-            return filename
-        except Exception as e:
-            self.logging.error('FileSaver: save_photo() error: ')
-            self.logging.exception(e)
-            pass
+        if(checkStorage() < 99):
+            filename = timestamp
+            filename = filename + ".jpg"
+            self.logging.info('saving file')
+            try:
+                cv2.imwrite(os.path.join(self.config["photos_path"], filename), image)
+                return filename
+            except Exception as e:
+                self.logging.error('FileSaver: save_photo() error: ')
+                self.logging.exception(e)
+                pass
+        else : 
+            self.logging.error('not enough space to save image')
+            return None
 
     def save_thumb(self,image,timestamp,format):
 
@@ -60,25 +64,64 @@ class FileSaver(Thread):
             self.logging.exception(e)
             pass
 
-
     def save_video(self, stream,timestamp):
         """
         Save raw video stream to disk
         :param stream: raw picamera stream object
         :return: none
         """
-        self.logging.info('FileSaver: Writing video...')
-        filename = timestamp
-        filenameMp4 = filename
-        filename = filename + ".h264"
-        filenameMp4 = filenameMp4 + ".mp4"
-        stream.copy_to(os.path.join(self.config["videos_path"], filename),seconds = self.config["video_duration_before_motion"] + self.config["video_duration_after_motion"])
-        self.logging.info('FileSaver: Done writing video ' + filename)
-        input_video = os.path.join(self.config["videos_path"], filename)
-        output_video = os.path.join(self.config["videos_path"], filenameMp4)
-        call(["MP4Box", "-fps", str(self.config["frame_rate"]), "-add", input_video, output_video])
-        os.remove(os.path.join(self.config["videos_path"], filename))
-        self.logging.info('Removed ' + filename)
+        if(checkStorage() < 99):
+            self.logging.info('FileSaver: Writing video...')
+            filename = timestamp
+            filenameMp4 = filename
+            filename = filename + ".h264"
+            filenameMp4 = filenameMp4 + ".mp4"
+            stream.copy_to(os.path.join(self.config["videos_path"], filename),seconds = self.config["video_duration_before_motion"] + self.config["video_duration_after_motion"])
+            self.logging.info('FileSaver: Done writing video ' + filename)
+            input_video = os.path.join(self.config["videos_path"], filename)
+            output_video = os.path.join(self.config["videos_path"], filenameMp4)
+            call(["MP4Box", "-fps", str(self.config["frame_rate"]), "-add", input_video, output_video])
+            os.remove(os.path.join(self.config["videos_path"], filename))
+            self.logging.info('Removed ' + filename)
+        else : 
+            self.logging.error('not enough space to save video')
+            return None
+    
+    def download_all_video():
+        timestamp = datetime.datetime.now()
+        filename = "video_"+timestamp.strftime('%Y-%m-%d-%H-%M-%S')
+        call(["cp", "-r", self.config["videos_path"], "data/"+filename])
+        call(["zip","-r","filename","data/"+filename])
+        call(["rm","-rf",filename])
+
+    def checkStorage():
+
+        # Disk information
+        description = getDfDescription()
+        disk_root = getDf()
+        out = disk_root[4].split('%')
+        return out[0]
+
+    def getDfDescription():
+        df = os.popen("df -h /")
+        i = 0
+        while True:
+            i = i + 1
+            line = df.readline()
+            if i==1:
+                return(line.split()[0:6])
+                                 
+    def getDf():
+        df = os.popen("df -h /")
+        i = 0
+        while True:
+            i = i + 1
+            line = df.readline()
+            if i==2:
+                return(line.split()[0:6])
+
+
+
